updated seeds.js
